requirements.txt
Flask
Flask-SQLAlchemy
Flask-Mail
pymysql
bcrypt
PyJWT